requirements.txt
ultralytics==8.3.0
torch==2.1.2
torchvision==0.16.2
pandas==2.1.4
pillow==10.2.0
matplotlib==3.8.2
scikit-learn==1.3.2
opencv-python==4.8.1.78
onnx==1.15.0
onnxruntime==1.16.3
